tasks/pos/requirements.txt
pydelphin
wn
networkx
matplotlib
toml
pygraphviz
